requirements.txt
python-dotenv
httpx
mcp[cli]